Ranges for the Theta histos are now doubles

+++ GentleKitty/Scripts/pLAnalysis/PLOnly.C
@@ -387,8 +387,8 @@ void FitPPVariations(const unsigned& NumIter, int imTBin, int system,
 	    const char* PhiHistoName;
 	    int PhiRebin = 1;
 	    int PhiConversion = 1; 
-	    int RangeProtonMin, RangeProtonMax;
-	    int RangeLambdaMin, RangeLambdaMax;
+	    double RangeProtonMin, RangeProtonMax;
+	    double RangeLambdaMin, RangeLambdaMax;
 	    if (iAngDist == 0) {
 	      PhiFile = "DimiPhi_pp_HM.root";
 	    PhiHistoName = "h_rkAngle_Mom2";
